test: remove duplicated tests
ConfigServiceTest has been migrated in Jupiter JUnit but it was not run
by Surefire because the name was not suffixed by Test.
Running all the tests in IntelliJ was running the file and I had to fix
some tests due to Mockito not behaving the same (Strictness is enabled
by default when running Mockito with JUnit Jupiter extension)
